Altun3s - Online Music Player Application
Using just HTML,CSS and Javascript, this application provides us with the option of inserting our own songs and playing them at our ease. Songs are listed with their duration as well as with their info. We can Play, Pause our current song as well as switch to the next or previous song. This application also gives us the option of playing any song we want in our library by simply clicking on the play button next to each song. This is similar to a Spotify Clone but is made using basic technologies. It is Responsive as well so songs can be easily played across various devices.
This Project also comes with a Status bar which helps us to see the duration of the music played as well as gives us the option to skip to our favourite parts!